added endpoint on interceptor creation
