+++ apps/ai-dial-admin/src/components/EntityMainProperties/SimpleEntityProperties.tsx
@@ -1,15 +1,15 @@
 import { FC, useCallback, useEffect, useState } from 'react';
 
+import CopyButton from '@/src/components/Common/CopyButton/CopyButton';
 import { TextInputField } from '@/src/components/Common/InputField/InputField';
 import TextAreaField from '@/src/components/Common/TextAreaField/TextAreaField';
-import { CreateI18nKey, RoutesI18nKey } from '@/src/constants/i18n';
+import { CreateI18nKey, EntitiesI18nKey, RoutesI18nKey } from '@/src/constants/i18n';
 import { useI18n } from '@/src/locales/client';
 import { DialBaseNamedEntity } from '@/src/models/dial/base-entity';
-import CopyButton from '@/src/components/Common/CopyButton/CopyButton';
-import { checkNameVersionCombination } from '@/src/utils/prompts/versions';
 import { DialRoute } from '@/src/models/dial/route';
-import { ApplicationRoute } from '@/src/types/routes';
 import { FieldError } from '@/src/models/error';
+import { ApplicationRoute } from '@/src/types/routes';
+import { checkNameVersionCombination } from '@/src/utils/prompts/versions';
 import { getErrorForDescription } from '@/src/utils/validation/description-error';
 import { getErrorForName } from '@/src/utils/validation/name-error';
 import { getErrorForPath } from '@/src/utils/validation/path-error';
@@ -85,6 +85,13 @@ const SimpleEntityProperties: FC<Props> = ({
     [entity, onChangeEntity, t],
   );
 
+  const onChangeEndpoint = useCallback(
+    (endpoint: string) => {
+      onChangeEntity({ ...entity, endpoint });
+    },
+    [entity, onChangeEntity],
+  );
+
   return (
     <div className="flex flex-col gap-6">
       <TextInputField
@@ -131,6 +138,16 @@ const SimpleEntityProperties: FC<Props> = ({
           errorText={pathError?.text}
           invalid={!!pathError}
           onChange={onChangePath}
+        />
+      )}
+
+      {view === ApplicationRoute.Interceptors && (
+        <TextInputField
+          elementId="endpoint"
+          fieldTitle={t(EntitiesI18nKey.Endpoint)}
+          placeholder={t(EntitiesI18nKey.EndpointPlaceholder)}
+          value={entity.endpoint}
+          onChange={onChangeEndpoint}
         />
       )}
     </div>
